refactor: extract tsconfig option and transform options merging logic (#8168)
Extracted tsconfig option and transform options merging logic.
