+++ crates/rolldown_plugin_vite_transform/src/utils.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use itertools::Either;
 use memchr::memmem;
 use oxc::{span::SourceType, transformer::TransformOptions};
-use rolldown_common::{JsxOptions, ModuleType};
+use rolldown_common::{LogWithoutPlugin, ModuleType, merge_tsconfig};
 use rolldown_plugin::SharedTransformPluginContext;
 use rolldown_utils::{pattern_filter::filter as pattern_filter, url::clean_url};
 
@@ -119,141 +119,26 @@ impl ViteTransformPlugin {
 
     if source_type.is_typescript() {
       let path = Path::new(cwd).join(id);
-      let tsconfig = self.resolver.find_tsconfig(path)?;
-
-      if let Some(tsconfig) = tsconfig {
+      if let Some(tsconfig) = self.resolver.find_tsconfig(path)? {
         // Tsconfig could be out of root, make sure it is watched
         let tsconfig_path = tsconfig.path.to_string_lossy();
         if !tsconfig_path.starts_with(cwd) {
           ctx.add_watch_file(&tsconfig_path);
         }
 
-        let compiler_options = &tsconfig.compiler_options;
-
-        // when both the normal options and tsconfig is set,
-        // we want to prioritize the normal options
-        if compiler_options.jsx.as_deref() == Some("preserve")
-          && transform_options
-            .jsx
-            .as_ref()
-            .is_none_or(|jsx| matches!(jsx, Either::Right(right) if right.runtime.is_none()))
-        {
-          transform_options.jsx = Some(Either::Left(String::from("preserve")));
+        let (merged_options, warnings) = merge_tsconfig(transform_options, &tsconfig, false);
+        for warning in warnings {
+          ctx.warn(LogWithoutPlugin {
+            message: warning.to_string(),
+            id: Some(id.to_string()),
+            code: Some(warning.kind().to_string()),
+            ..Default::default()
+          });
         }
-        if !matches!(&transform_options.jsx, Some(Either::Left(left)) if left == "preserve") {
-          let mut jsx = if let Some(Either::Right(jsx)) = transform_options.jsx {
-            jsx
-          } else {
-            JsxOptions::default()
-          };
-
-          if compiler_options.jsx_factory.is_some() && jsx.pragma.is_none() {
-            jsx.pragma.clone_from(&compiler_options.jsx_factory);
-          }
-          if compiler_options.jsx_import_source.is_some() && jsx.import_source.is_none() {
-            jsx.import_source.clone_from(&compiler_options.jsx_import_source);
-          }
-          if compiler_options.jsx_fragment_factory.is_some() && jsx.pragma_frag.is_none() {
-            jsx.pragma_frag.clone_from(&compiler_options.jsx_fragment_factory);
-          }
-
-          if jsx.runtime.is_none() {
-            match compiler_options.jsx.as_deref() {
-              Some("react") => {
-                jsx.runtime = Some(String::from("classic"));
-                // this option should not be set when using classic runtime
-                jsx.import_source = None;
-              }
-              Some("react-jsx") => {
-                jsx.runtime = Some(String::from("automatic"));
-                // these options should not be set when using automatic runtime
-                jsx.pragma = None;
-                jsx.pragma_frag = None;
-              }
-              Some("react-jsxdev") => jsx.development = Some(true),
-              _ => {}
-            }
-          }
-          transform_options.jsx = Some(Either::Right(jsx));
-        }
-
-        if transform_options.decorator.as_ref().is_none_or(|decorator| decorator.legacy.is_none()) {
-          let mut decorator = transform_options.decorator.unwrap_or_default();
-
-          if compiler_options.experimental_decorators.is_some() {
-            decorator.legacy = compiler_options.experimental_decorators;
-          }
-
-          if compiler_options.emit_decorator_metadata.is_some() {
-            decorator.emit_decorator_metadata = compiler_options.emit_decorator_metadata;
-          }
-
-          transform_options.decorator = Some(decorator);
-        }
-
-        // | preserveValueImports | importsNotUsedAsValues | verbatimModuleSyntax | onlyRemoveTypeImports |
-        // | -------------------- | ---------------------- | -------------------- |---------------------- |
-        // | false                | remove                 | false                | false                 |
-        // | false                | preserve, error        | -                    | -                     |
-        // | true                 | remove                 | -                    | -                     |
-        // | true                 | preserve, error        | true                 | true                  |
-        let mut typescript = transform_options.typescript.unwrap_or_default();
-        if typescript.only_remove_type_imports.is_none() {
-          if compiler_options.verbatim_module_syntax.is_some() {
-            typescript.only_remove_type_imports = compiler_options.verbatim_module_syntax;
-          } else if compiler_options.preserve_value_imports.is_some()
-            || compiler_options.imports_not_used_as_values.is_some()
-          {
-            let preserve_value_imports = compiler_options.preserve_value_imports.unwrap_or(false);
-            let imports_not_used_as_values =
-              compiler_options.imports_not_used_as_values.as_deref().unwrap_or("remove");
-            typescript.only_remove_type_imports = if !preserve_value_imports
-              && imports_not_used_as_values == "remove"
-            {
-              Some(true)
-            } else if preserve_value_imports
-              && (imports_not_used_as_values == "preserve" || imports_not_used_as_values == "error")
-            {
-              Some(false)
-            } else {
-              // warnings.push(
-              //   `preserveValueImports=${preserveValueImports} + importsNotUsedAsValues=${importsNotUsedAsValues} is not supported by oxc.` +
-              //     'Please migrate to the new verbatimModuleSyntax option.',
-              // )
-              Some(false)
-            };
-          }
-        }
-
-        let disable_use_define_for_class_fields = !compiler_options
-          .use_define_for_class_fields
-          .unwrap_or_else(|| is_use_define_for_class_fields(compiler_options.target.as_deref()));
-
-        let mut assumptions = transform_options.assumptions.unwrap_or_default();
-        assumptions.set_public_class_fields = Some(disable_use_define_for_class_fields);
-        typescript.remove_class_fields_without_initializer =
-          Some(disable_use_define_for_class_fields);
-
-        transform_options.typescript = Some(typescript);
-        transform_options.assumptions = Some(assumptions);
+        transform_options = merged_options;
       }
     }
 
     Ok((source_type, transform_options.try_into().map_err(|err: String| anyhow::anyhow!(err))?))
   }
-}
-
-fn is_use_define_for_class_fields(target: Option<&str>) -> bool {
-  let Some(target) = target else { return false };
-
-  if target.len() < 3 || !&target[..2].eq_ignore_ascii_case("es") {
-    return false;
-  }
-
-  let reset = &target[2..];
-  if reset.eq_ignore_ascii_case("next") {
-    return true;
-  }
-
-  reset.parse::<usize>().is_ok_and(|x| x > 2021)
 }
